fix: open the dashboard page from top (scroll)
